Comit token refresh when opp searching is  started

+++ server/src/websockets/handlers/searchPlayerHandler.ts
@@ -1,9 +1,16 @@
 import { Server, Socket } from "socket.io";
 import SearchUserService from "../../services/searchService.js";
 import SearchUserData from "../../types/searchUserData.js";
+import TokenService from "../../services/tokenService.js";
 
 export default function registerSearchPlayerHandlers(io: Server, socket: Socket) {
     socket.on("search opponent:searching", async ({username, trophies}: Omit<SearchUserData, "socketId">) => {
+        const token = socket.handshake.headers.authorization as string;
+        if(!TokenService.validateAccessToken(token)){
+            socket.emit("search opponent:invalid token");
+            return;
+        }   
+
         const opponent = await SearchUserService.findSuitableOpponent(trophies);
         if (opponent) {
             await SearchUserService.removeSearchUser(opponent.socketId);//нужно еще будет учитывать, что пользователь мог ливнуть, когда искал оппонента
@@ -21,7 +28,4 @@ export default function registerSearchPlayerHandlers(io: Server, socket: Socket)
 
         await SearchUserService.addSearchUser({ socketId: socket.id, trophies, username });
     });
-
-
-
 }
